analytics/requirements.txt
agate==1.6.1
alabaster==0.7.11
alembic==1.2.1
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
apache-airflow==1.10.5
apispec==3.0.0
appdirs==1.4.3
asn1crypto==1.3.0
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
azure-common==1.1.23
azure-core==1.9.0
azure-storage-blob==12.6.0
azure-storage-common==2.1.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.2.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
blis==0.4.1
bokeh==0.13.0
boto==2.49.0
boto3==1.9.176
botocore==1.12.176
Bottleneck==1.2.1
CacheControl==0.12.6
cached-property==1.5.1
cachetools==3.1.1
Camelot==12.6.29
catalogue==1.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
colorlog==4.0.2
conda==4.5.11
conda-build==3.15.1
configparser==3.5.3
constantly==15.1.0
contextlib2==0.5.5
croniter==0.3.30
cryptography==3.3.1
cycler==0.10.0
cymem==2.0.3
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
datashape==0.5.4
dbt==0.18.1
dbt-bigquery==0.18.1
dbt-core==0.18.1
dbt-postgres==0.18.1
dbt-redshift==0.18.1
dbt-snowflake==0.18.1
decorator==4.3.0
defusedxml==0.5.0
descartes==1.1.0
dill==0.2.9
distributed==1.23.1
distro==1.5.0
docker==4.4.1
docker-compose==1.28.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dumb-init==1.2.2
Elixir==0.7.1
en-core-web-sm==2.2.5
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.8
Fiona==1.8.6
firebase-admin==4.5.1
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Cors==3.0.6
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
funcy==1.14
future==0.16.0
gensim==3.8.3
geopandas==0.5.0
gevent==1.3.6
glob2==0.6
gmpy2==2.0.8
google-api-core==1.16.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==1.25.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.6.0
GoogleNews==1.3.8
googletrans==2.4.0
greenlet==0.4.15
grpcio==1.35.0
gunicorn==19.9.0
h5py==2.8.0
heapdict==1.0.0
hologram==0.0.10
html5lib==1.0.1
httplib2==0.19.0
hyperlink==18.0.0
idna==2.7
ijson==2.4
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
iso8601==0.1.12
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.13.2
json-merge-patch==0.2
json-rpc==1.12.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
keyring==13.2.1
kiwisolver==1.0.1
langdetect==1.0.8
lazy-object-proxy==1.3.1
leather==0.3.3
llvmlite==0.24.0
locket==0.2.0
lockfile==0.12.2
Logbook==1.5.3
lxml==4.2.5
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.19.0
matplotlib==2.2.3
mccabe==0.6.1
minimal-snowplow-tracker==0.0.2
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
monkeylearn==3.5.2
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
msrest==0.6.19
multipledispatch==0.6.0
munch==2.3.2
murmurhash==1.0.2
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.4
newsapi-python==0.2.6
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.16.3
numpydoc==0.8.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.6
opsgenie-sdk==2.0.3
ordereddict==1.1
oscrypto==1.2.0
packaging==17.1
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.7.2
parsedatetime==2.4
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pbr==5.4.4
pendulum==1.4.4
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
plac==1.1.3
plotly==4.1.1
pluggy==0.7.1
ply==3.11
preprocessor==1.1.3
preshed==3.0.2
prison==0.1.0
progressbar==2.5
prometheus-client==0.3.1
prompt-toolkit==1.0.15
proto-plus==1.13.0
protobuf==3.8.0
psutil==5.4.7
psycopg2==2.8.4
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.8.2
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
PyJWT==1.7.1
pyLDAvis==2.1.2
pylint==2.1.1
pymongo==3.10.1
PyNaCl==1.4.0
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.4.7
pyproj==2.1.3
pyrsistent==0.15.4
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-logbook==1.2.0
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-daemon==2.1.2
python-dateutil==2.8.0
python-dotenv==0.15.0
python-editor==1.0.4
python-slugify==3.0.2
python3-openid==3.1.0
pytimeparse==1.1.8
pytz==2019.1
pytzdata==2019.3
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.3.0
scorecardpy==0.1.7.4
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
service-identity==17.0.0
setproctitle==1.1.10
Shapely==1.6.4.post2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smart-open==2.0.0
snowballstemmer==1.2.1
snowflake-connector-python==2.2.10
sortedcollections==1.0.1
sortedcontainers==2.0.5
spacy==2.2.4
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.2.3
srsly==1.0.2
stanza==1.0.1
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
tabulate==0.8.5
tblib==1.3.2
Tempita==0.5.2
tenacity==6.2.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
textblob==0.15.3
texttable==1.6.3
thinc==7.4.0
thrift==0.11.0
toolz==0.9.0
torch==1.5.0
tornado==5.1
tqdm==4.46.0
traitlets==4.3.2
tweepy==3.8.0
Twisted==18.7.0
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.23
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.15.4
widgetsnbextension==3.4.1
wrapt==1.10.11
WTForms==2.2.1
xlrd==0.7.1
XlsxWriter==1.1.0
xlwt==0.7.2
zict==0.1.3
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==4.5.0
